C:\Users\rdboylan\Documents\Kornak\bifs\bifs\src>py bifs_gui.py Completely new and more efficient computation of mean and sd.  Used for production run to get PET prior distn. I only saved the k_mod instead of the whole bifs object. I used a running mean/sd algorithm to get those values far more efficiently. With these changes and shutting down all other large applications, it was just barely able to run in the 32G on this machine.  Profiling results appear below.
The code in this commit isn't quite what ran.  I think the only difference is the code here
prints how many images were processed on the console.  I neglected to save that change before
starting the run.  There were ~ 164 directories, and the stats below suggest 154 had images that
were processed.

The code to get the mean and sd is easily adaptable to incremental approaches.
The obvious next optimization is to dispense with holding the complete list of moduli for all
images.  This should result in huge memory savings.

I think much of the time in _statsImages was actually writing out the file with the results.

Also removed the test limit to 10 images.

Skipping p10933_B10-272_2010-09-22
Skipping p10933_B10-272_2011-03-14
D_Lib: debug printing for Skipping p10933_B10-272_2012-10-15
files [.*] and level [100] is turned on
D_Lib: debug printing for files [.*] and level [200] is turned on
D_Lib: debug printing for files [.*] and level [300] is turned on
12560:vf_shex.cpp(88): INFO: DllCanUnloadNow returned S_OK.
         293385 function calls (290430 primitive calls) in 854.111 seconds

   Ordered by: cumulative time
   List reduced from 907 to 15 due to restriction <15>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1   10.450   10.450  813.056  813.056 bifs_gui.py:123(_scanImages)
      154    0.002    0.000  799.635    5.192 bifs.py:313(load_image_file)
      154   25.851    0.168  434.808    2.823 bifs.py:365(load_image)
      154    0.711    0.005  351.754    2.284 dataobj_images.py:208(get_fdata)
      154  350.863    2.278  350.864    2.278 {method 'astype' of 'numpy.ndarray' objects}
      154    0.000    0.000  195.409    1.269 basic.py:566(fftn)
      154    1.479    0.010  195.409    1.269 basic.py:627(_raw_fftn_dispatch)
      154  152.567    0.991  193.927    1.259 basic.py:527(_raw_fftnd)
      154    6.967    0.045  117.573    0.763 bifs.py:463(set_prior_from_param_func)
      154  110.604    0.718  110.604    0.718 fourier.py:153(ixsc)
     1388    0.004    0.000   62.292    0.045 {built-in method numpy.core._multiarray_umath.implement_array_function}
      154    0.000    0.000   61.738    0.401 <__array_function__ internals>:2(angle)
      154   61.736    0.401   61.737    0.401 function_base.py:1419(angle)
      462   41.289    0.089   41.339    0.089 basic.py:148(_fix_shape)
        1   40.516   40.516   41.056   41.056 bifs_gui.py:193(_statsImages)