Changes to try to get import actually working as a package. Running something from the examples directory that attempted to import bifs revealed the existing structure didn't work.
These changes BREAK running bifs_gui from the command line in place,
as has been my practice.  When bifs_gui imports bifs, it is more like an include of the single file bifs.py.  During the import their is no defined module space, as a result of which the relative imports this commit installs in bifs fail.