Our first test, and first test failure.  test_reset.py named because it was going to catch problems if we reset the image or the parameters.  However, it fails without any resetting at all!
Test environment
C:\Users\rdboylan\Documents\Kornak\bifs>set PYTHONPATH=C:\Users\rdboylan\Documents\Kornak\bifs
I.e., in the top of the project directory, set PYTHONPATH to tht directory.  This is so import bifs will work in the test code.
Then execute pytest in that directory.

This particular discrepancy may not be all that important, though it does result in the intended high variance at [0, 0] not being picked up.  But it also shows a general problem with the sequencing of the commands.