kdist and k_image becomes function on bifs, which will require client changes.  They move toward lazy evaluation as a way of avoiding some sequencing issues about which operations must be done in which order.
FunctionalPrior gets a function given the list of allowed function types.  bifs now uses it.  Probably bifs shouldn't hold the value at all, but bifs_gui is counting on it.

example02 in presentation is now complete and results sent to John.

The use of functions to access some, but not all, values in bifs is inconsistent and confusing.

This remains a work in progress.

Note the example code has to peek inside the prior to get the scale value; that needs proper accessors too.