Continued work on import.  Adjust to BIFS as package name. Correct comment about use of multiprocessing; we aren't using it now. Still not working.
C:\Users\rdboylan\Documents\Kornak>py bifs_gui.py
D_Lib: debug printing for files [.*] and level [100] is turned on
D_Lib: debug printing for files [.*] and level [200] is turned on
D_Lib: debug printing for files [.*] and level [300] is turned on
12972:vf_shex.cpp(88): INFO: DllCanUnloadNow returned S_OK.
Traceback (most recent call last):
  File "bifs_gui.py", line 105, in getEmpiricalPrior
    if self._scanImages(topDir):
  File "bifs_gui.py", line 152, in _scanImages
    self._statsAccumulate(b.mod_image)
AttributeError: 'bifs' object has no attribute 'mod_image'

The problem is that the code in bifs_gui is making assumptions about the
internals of the bifs object that are no longer valid, and that I haven't
provided proper accessors to deal with it.

This also highlights that much of the logic in bifs_gui for empirical priors
really belongs closer to the bifs class.  But note that since the operation
is on a set of images, it probably should not go entirely in the bifs class,
which represents a single image.