Add EmpiricalScanner class; first working presentation03 with rescaling.
EmpiricalScanner takes much of the logic currently in bifs_gui for
building up empirical distributions from scans.  It also adds the ability
to get the distribution of voxels in image space for the scanned files.
Currently I've disabled the distribution stuff for speed of development.

The voxel distribution if a subsample of all voxels; subsampling uses
the new random number scheme in numpy 1.17; the requirements have been
modified (not sure the syntax is right) to reflect that.

The presentation now changes the distribution of voxel intensities in the
MRI to match that of a typical PET.  The revealed so far unnoticed interference patterns
in the background.

presentation03b.pdf has the results; sent them to John.  The file is ot in git.

The bifs transformation still isn't making much difference.