Latest tweaks to outputs. JK request: Could we add a couple of intermediate Inverse Power Decay reconstructions between scale 5e-05 and 0.0005?
Also for the linear decay try a couple of scales that go even smaller – e.g. scale 5e-06 and 5e-07 and remove 0.005

I think we can drop the banded priors for the moment. I don’t think they are adding anything and we need to have an idea as to where we want to accentuate frequencies in order to gain anything from that.