investigate09 looks at whole image unmasked. Also looks at sum of squares, which are supposed to be maintained.  And they seem to be (across different phase values from the same prior). Shared results with JK.
This is whole image enough to rule out the possibility that the image distribution is changing because of different areas rotating in and out of the masked area.  And it shows clearly that neither the mean nor the max are maintained across different phases.  Phase 0 produces a few really high values.

OTOH, the reliance on ep3 and vox3 is a reliance on the masked values only.  In particular, the transform I'm doing mamps everything, included the background region, into the voxel intensities in the focal area only.  That doesn't make a lot of sense.