EmpiricalScanner gains ability to accumumulate info on phase and correlation phase/modulus. This changes the interface, though backward compatibility is provided for now.
empirical_scan switched to new interace and asks for the new info, which it puts in a separate file,
ep3-phase.npz.

This runs and the correlations are  in the correct range.
No correlation is available for the 0, 0, 0 cell (value is na) since phase is always 0.  This throws an error during the run,
but does not block the other calculations.

NEED AN INDEPENDENT TEST TO VERIFY CALCULATIONS ARE CORRECT.

From a separate python session run after:
>>> dphs = np.load(PHASENAME)
>>> corr=dphs["corr"]
>>> corr[0:4, 0:4, 0]
array([[        nan, -0.09266272, -0.09676847,  0.13431336],
       [ 0.23990765, -0.00198639, -0.10975952,  0.03687923],
       [ 0.12617723, -0.08565475,  0.17635505, -0.15894834],
       [ 0.01928849,  0.12259451, -0.08748215,  0.333021  ]])
>>> np.nanpercentile(corr.ravel(), range(0, 101, 10))
array([-8.57153255e-01, -2.17301718e-01, -1.44983250e-01, -9.11219839e-02,
       -4.41747342e-02, -6.80926792e-14,  4.41747342e-02,  9.11219839e-02,
        1.44983250e-01,  2.17301718e-01,  8.57153255e-01])
That distribution i s very symmetric, but because of Hermitian symmetry it probably has to be.