Fixes#24: error on exit Also fix minor typos in ToDo.otl and note progress there.
The immediate cause of the error was that the app variable was unknown in the scope of the close method of MainWindow.

But it didn't make sense to be exec'ing the app from the close code anyway.  It has already been exec'd, and so another exec is a no-op (or so says the internet).

I have made the  appllication an instance variable of MainWindow so the latter's close() method has access to it.  This is NOT based on any examples or advice, and so may not be the recommended style.  close() simply call's the app's quit() method (which exits with status 0).

PyQt5 and Python 3 have brought considerable changes to the preferred style of exiting, reflected here:
1. sys.exit() is unnecessary; app.exit() or quit() will exit, passing any status codes outward.
2. Use of exec_ is no longer necessary since exec is not a keywork in Python3.

See https://stackoverflow.com/questions/8026101/correct-way-to-quit-a-qt-program and
https://stackoverflow.com/questions/22610720/should-i-use-app-exec-or-app-exec-in-my-pyqt-application

It continues to be the case that exiting via the exit menu entry on the BIFS operations menu goes through MainWindow.close() while hitting the X to close the window does not.  This seems undesirable, but both work and do the same thing.

Note the following, which might be relevant in the future, from the Qt docs:
We recommend that you connect clean-up code to the aboutToQuit() signal, instead of putting it in your application's main() function. This is because, on some platforms, the QApplication::exec() call may not return.
At the moment I don't do anything but exit.