Progress getting examples to run 1. Change paths for image loads.  But this makes assumptions about where it is run from, namely the directory with the example code. 2. Many *_image values that used to be direct access are now by function call. 3. import path changed for util.  It is now a bit ridiculous.
1D already ran, having changes incorporated earlier.
The 2D code now runs, with the previous fix to util.
The 3D code fails, even with these changes.

The 3D example reads its input as 2D and then gets index out of range when line
# Look at the prior, liklelihood, and posterior at a voxel
tries to to reference the 3rd dimension with shape[2].

Finally one of these files triggered this warning, which is unresolved:
  | j:\source\repos\bifs\bifs\bifs_util\util.py:210: MatplotlibDeprecationWarning: Calling gca() with keyword arguments was deprecated in Matplotlib 3.4. Starting two minor releases later, gca() will take no keyword arguments. The gca() function should only be used to get the current axes, or if no axes exist, create new axes with default keyword arguments. To create a new axes with non-default arguments, use plt.axes() or plt.subplot().
  |   ax = fig.gca(projection='3d')

The offending line is in util.py:210 and was triggered by the 1D example, and probably 3D if it gets far enough.